# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Environment:
    properties:
      createTime:
        description: 创建时间
        type: integer
      desc:
        description: 环境描述
        type: string
      etcdEndpoints:
        description: ETCD地址
        items:
          type: string
        type: array
      id:
        description: Id
        type: string
      name:
        description: 环境名称
        type: string
    required:
    - etcdEndpoints
    - name
    type: object
  models.Export:
    properties:
      address:
        description: 节点地址
        type: string
      createTime:
        description: 创建时间
        type: integer
      env:
        description: 环境名称
        type: string
      id:
        description: Id
        type: string
      targets:
        description: 节点地址列表
        items:
          type: string
        type: array
      type:
        description: 节点类型
        type: string
    required:
    - address
    - type
    type: object
  models.PromJob:
    properties:
      configsType:
        description: 配置类型：static || http
        type: string
      createTime:
        description: 创建时间
        type: integer
      id:
        description: Id
        type: string
      jobName:
        description: 采集任务名称
        type: string
      metricsPath:
        description: Metrics Path
        type: string
      scheme:
        description: 协议 Scheme
        type: string
      targets:
        description: static 采集目标
        items:
          type: string
        type: array
      url:
        description: http 请求地址
        type: string
    required:
    - configsType
    - jobName
    - metricsPath
    - scheme
    type: object
...
  resp.Response:
    properties:
      code:
        description: 错误码 状态码不为200时返回
        type: integer
      data:
        description: 返回数据
        type: object
      message:
        description: 错误信息 状态码不为200时返回
        type: string
      more_info:
        description: 错误扩展信息 状态码不为200时返回
        type: object
      status:
        description: 状态码 200:成功   其他失败
        type: integer
      type:
        description: 错误类型 状态码不为200时返回
        type: string
    type: object
info:
  contact:
    email: leizhou.lin@watrix.ai
    name: API Support
    url: github.com/souliot
  description: Golang api of demo
  termsOfService: github.com/souliot
  title: DevOps 开发文档
  version: v1.0.0
paths:
  /env:
    get:
      consumes:
      - application/json
      description: 获取环境列表
      parameters:
      - description: node type
        in: query
        name: type
        type: string
      - description: node address
        in: query
        name: address
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 []Environment
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Environment'
                  type: array
              type: object
      summary: 获取环境列表
      tags:
      - 环境
    post:
      consumes:
      - application/json
      description: 添加环境的接口
      parameters:
      - description: 环境节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.Environment'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Environment
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Environment'
              type: object
      summary: 添加环境
      tags:
      - 环境
    put:
      consumes:
      - application/json
      description: 修改环境的接口
      parameters:
      - description: 环境节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.Environment'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Environment
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Environment'
              type: object
      summary: 修改环境
      tags:
      - 环境
  /env/{id}:
    delete:
      consumes:
      - application/json
      description: 删除环境的接口
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Environment
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Environment'
              type: object
      summary: 删除环境
      tags:
      - 环境
    get:
      consumes:
      - application/json
      description: 获取环境信息
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Environment
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Environment'
              type: object
      summary: 获取环境信息
      tags:
      - 环境
  /export:
    get:
      consumes:
      - application/json
      description: 获取 Prometheus http_sd_config 的接口
      parameters:
      - description: node type
        in: query
        name: type
        type: string
      - description: node address
        in: query
        name: address
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 []Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Export'
                  type: array
              type: object
      summary: 获取监控节点列表
      tags:
      - 监控
    post:
      consumes:
      - application/json
      description: 添加 Prometheus http_sd_config 的接口
      parameters:
      - description: 监控节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.Export'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Export'
              type: object
      summary: 添加监控节点
      tags:
      - 监控
    put:
      consumes:
      - application/json
      description: 修改 Prometheus http_sd_config 的接口
      parameters:
      - description: 监控节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.Export'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Export'
              type: object
      summary: 修改监控节点
      tags:
      - 监控
  /export/{id}:
    delete:
      consumes:
      - application/json
      description: 删除 Prometheus http_sd_config 的接口
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Export'
              type: object
      summary: 删除监控节点
      tags:
      - 监控
    get:
      consumes:
      - application/json
      description: 获取 Prometheus http_sd_config 的接口
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Export'
              type: object
      summary: 获取监控节点信息
      tags:
      - 监控
  /export/type/{type}:
    get:
      consumes:
      - application/json
      description: 获取 Prometheus http_sd_config 的接口
      parameters:
      - description: node type
        in: path
        name: type
        required: true
        type: string
      - description: node env
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 Export
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Export'
              type: object
      summary: 获取监控节点
      tags:
      - Prom
  /prom/conf:
    get:
      consumes:
      - application/json
      description: 生成Prom配置文件
      parameters:
      - description: 生成配置文件路径
        in: query
        name: path
        type: string
      - description: 抓取时间间隔
        in: query
        name: scrapeInterval
        type: string
      - description: 报警检测时间间隔
        in: query
        name: evaluationInterval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 生成Prom配置文件
      tags:
      - Prom
...
  /promjob:
    get:
      consumes:
      - application/json
      description: 获取任务列表
      parameters:
      - description: node address
        in: query
        name: jobName
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 []PromJob
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PromJob'
                  type: array
              type: object
      summary: 获取任务列表
      tags:
      - 任务
    post:
      consumes:
      - application/json
      description: 添加任务的接口
      parameters:
      - description: 任务节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.PromJob'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 PromJob
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PromJob'
              type: object
      summary: 添加任务
      tags:
      - 任务
    put:
      consumes:
      - application/json
      description: 修改任务的接口
      parameters:
      - description: 任务节点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.PromJob'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 PromJob
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PromJob'
              type: object
      summary: 修改任务
      tags:
      - 任务
  /promjob/{id}:
    delete:
      consumes:
      - application/json
      description: 删除任务的接口
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 PromJob
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PromJob'
              type: object
      summary: 删除任务
      tags:
      - 任务
    get:
      consumes:
      - application/json
      description: 获取任务信息
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 PromJob
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PromJob'
              type: object
      summary: 获取任务信息
      tags:
      - 任务
...
swagger: "2.0"